Use MUnit in RouteAsUriTemplateSuite
